Refactoring sendController
