+++ src/utils/getWeb3.js
@@ -1,4 +1,3 @@
-require('dotenv').config()
 const HDWalletProvider = require('@truffle/hdwallet-provider')
 const Web3 = require('web3')
 const rpc = process.env.RPC
